app/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudinary==1.41.0
constantly==23.10.4
crashtest==0.4.1
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.3
dnspython==2.7.0
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
Faker==30.6.0
fastapi==0.115.5
fastapi-limiter==0.1.6
fastapi-mail==1.4.2
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
installer==0.7.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==24.3.1
lxml==5.3.0
MarkupSafe==2.1.5
mongoengine==0.29.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
parsel==1.9.1
passlib==1.7.4
pexpect==4.9.0
pika==1.3.2
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
Protego==0.3.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyDispatcher==2.0.7
pymongo==4.10.1
pyOpenSSL==24.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
queuelib==1.7.0
RapidFuzz==3.10.0
redis==5.2.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
Scrapy==2.11.2
service-identity==24.1.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.41.3
tldextract==5.1.2
tomlkit==0.13.2
trove-classifiers==2024.9.12
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.26.6
w3lib==2.2.1
Werkzeug==3.0.4
xattr==1.1.0
zope.interface==7.1.0
